backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.3",
]
